fix: Avoid duplicate active menu entries
If the menu contains
1) a link to a contenttype listing and
2) a link to a specific record of this very contenttype
both menu entries were marked as active.
This is now fixed, only the link to the very record is active.

The issue appeared bc we want to highlight 1)
if any record of this contenttype is opened.
E.g. if a blogpost is open, and there is a menu entry for the
blog-contenttype-listing we want it to be highlighted.